SAFE/requirements.txt
google-genai
python-dotenv
sounddevice
soundfile
numpy
pyttsx3
pandas
rapidfuzz